+++ test/component/test.component.temp_directory/entry.cpp
@@ -53,7 +53,7 @@ int main(int argc, char* argv[])
 
     XTESTS_COMMANDLINE_PARSEVERBOSITY(argc, argv, &verbosity);
 
-    if (XTESTS_START_RUNNER("test.component.temp_directory", verbosity))
+    if (XTESTS_START_RUNNER(__FILE__, verbosity))
     {
         XTESTS_RUN_CASE(test_None);
         XTESTS_RUN_CASE(test_EmptyOnOpen);